+++ packages/bridge/src/runtime/composables.ts
@@ -9,13 +9,13 @@ import { defu } from 'defu'
 import { useRouter as useVueRouter, useRoute as useVueRoute } from 'vue-router/composables'
 import { hasProtocol, joinURL, parseURL } from 'ufo'
 import { useNuxtApp, callWithNuxt } from './nuxt'
-import { createError, showError } from './error'
-import type { NuxtError } from './error'
+import { createError, showError } from './composables/error'
+import type { NuxtError } from './composables/error'
 
 export { useLazyAsyncData, refreshNuxtData } from './composables/asyncData'
 export { useLazyFetch } from './composables/fetch'
 export { useCookie } from './cookie'
-export { clearError, createError, isNuxtError, throwError, showError, useError } from './error'
+export { clearError, createError, isNuxtError, throwError, showError, useError } from './composables/error'
 export { useRequestHeaders, useRequestEvent } from './ssr'
 
 export * from 'vue'
